game: Fix board not being up-to-date.
When a player leaves a live game and reenters, we
need to fetch the most updated board state and render
it.
Because of recoil & a bug in singleton pattern a redundant
board state was getting rendered.

We no longer fetch individual game state from recoil and also
update the singleton cache everytime the live game component
mounts.